re-organize functions in relation to MITgcmTools.jl - mv readbin, readncfile, findidx, readncdata, readdata to src/utilities.jl - mv parsemeta, readAvailDiagnosticsLog to MITgcmTools.jl - rm read_nctiles (newer version already in MITgcmTools.jl)
