+++ src/writenctiles.jl
@@ -145,27 +145,6 @@ replacevalues(vals,td::TileData) = TileData(vals,td.tileinfo,td.tilesize,td.prec
 
 
 """
-    findidx(A,val)
-
-Helper function for getting the indices for tiles. A is a gcmfaces struct, val is a
-    numeric value to get the indices of. Returns a Dict of the indices for each face.
-    Not currently exported. Maybe move to MeshArrays?
-"""
-function findidx(A,val)
-    tileidx = Dict()
-    if isa(A,MeshArrays.gcmfaces)
-        nfaces = A.nFaces
-    elseif isa(A,MeshArray)
-        nfaces = A.grid.nFaces
-    end
-    for i = 1:nfaces
-        idx = findall(x->x==val,A.f[i])
-        tileidx[i] = [x for x in idx]
-    end
-    return tileidx
-end
-
-"""
     gettile(fldvals,tileinfo,tilesize,tilenum::Int)
 
 Helper function for retrieving a tile from a gcmfaces struct as a numeric Array. Not
@@ -267,107 +246,6 @@ Retrieve the ith time step from the NetCDF file.
 function getindex(d::NCData,i) # Gets data at time index i
     readncdata(d,i)
 end
-
-"""
-    readbin(fname::String,prec::Type,iosize::Tuple,fldidx=1)
-
-Read in a binary file as an array as previously done via `MeshArrays.read_bin`.  Not exported.
-"""
-function readbin(fname::String,prec::Type,iosize::Tuple,fldidx=1)
-    n3 = 1
-    if length(iosize) == 3
-        n1,n2,n3 = iosize
-    else
-        n1,n2 = iosize
-    end
-
-    if prec == Float64
-        reclen = 8
-    else
-        reclen = 4
-    end
-
-    #if isempty(n3)
-    #    n3=Int64(stat(fname).size/n1/n2/reclen)
-    #end
-
-    field = Array{prec,1}(undef,(n1*n2*n3))
-    fid = open(fname)
-    if fldidx != 1
-        seek(fid,n1*n2*n3*(fldidx-1)*reclen)
-    end
-    read!(fid,field)
-    close(fid)
-    field = hton.(field)
-
-    n3>1 ? s=(n1,n2,n3) : s=(n1,n2)
-    return reshape(field,s)
-
-end
-
-"""
-    readbin(flddata::BinData,tidx=1)
-
-Read in a binary file as an array as previously done via `MeshArrays.read_bin`
-"""
-readbin(flddata::BinData,tidx=1) = readbin(flddata.fnames[tidx],flddata.precision,flddata.iosize,flddata.fldidx)
-"""
-    readncdata(var::NCData,i::Union{Colon,Integer}=:)
-
-Read netcdf file as specified in `NCData` argument. Optional
-argument `i` can be used to read a specific records / times. Not exported.
-"""
-function readncdata(var::NCData,i::Union{Colon,Integer}=:)
-    ds = Dataset(var.fname)
-    ndims = length(size(ds[var.varname]))
-
-    if ndims == 1
-        values = ds[var.varname][i]
-    elseif ndims == 2
-        values = ds[var.varname][:,i]
-    elseif ndims == 3
-        values = ds[var.varname][:,:,i]
-    elseif ndims == 4
-        values = ds[var.varname][:,:,:,i]
-    else
-        println("Variables greater than four dimensions not currently supported")
-    end
-    close(ds)
-    return values
-end
-
-"""
-    readdata(flddata,tidx=1)
-
-Generic wrapper function to read data from a file regardless of file/data type. Not exported.
-"""
-function readdata(flddata,tidx=1)
-    if isa(flddata,BinData)
-        res = readbin(flddata,tidx)
-    elseif isa(flddata,NCData)
-        res = readncdata(flddata,tidx)
-    elseif isa(flddata,TileData)
-        if isa(flddata.vals,BinData)
-            iosize = flddata.vals.iosize
-            grid = flddata.tileinfo["XC"].grid
-            if length(iosize) == 3
-                f = Array{Array{flddata.vals.precision,2},2}(undef,grid.nFaces,iosize[3])
-                exarray = MeshArray(grid,f)
-            else
-                exarray = flddata.tileinfo["XC"]
-            end
-            res = read(readbin(flddata.vals,tidx),exarray)
-        else
-            if isa(flddata.vals,MeshArray) || isa(flddata.vals,MeshArrays.gcmfaces)
-                res = flddata.vals
-            else
-                res = flddata.vals[tidx]
-            end
-        end
-    end
-    return res
-end
-
 
 """
     addDim(ds::NCDatasets.Dataset,dimvar::NCvar) # NCDatasets
@@ -803,68 +681,4 @@ Helper function: determines whether an array of dimensions has a time dimension.
 """
 function hastimedim(dims::Array{NCvar})
     return any(istimedim.(dims))
-end
-
-"""
-    parsemeta(metafile)
-
-Parse out an `MITgcm` metadata file and return a `Dict` of fields in the file.
-"""
-function parsemeta(metafile)
-
-    meta = read(metafile,String)
-    meta = split(meta,";\n")
-    meta = meta[isempty.(meta).==false]
-    meta = replace.(meta,Ref(",\n"=>";"))
-    meta = replace.(meta,Ref("\n"=>""))
-    meta = replace.(meta,Ref("}"=>"]"))
-    meta = replace.(meta,Ref("{"=>"["))
-    #meta = replace.(meta,Ref(" "=>""))
-    meta = replace.(meta,Ref("'"=>"\""))
-    meta = replace.(meta,Ref(";]"=>"]"))
-    meta = replace.(meta,Ref(","=>" "))
-
-    meta = split.(meta,"=")
-    meta = [[replace(x[1]," "=>"") x[2]] for x in meta]
-
-    metaDict = Dict{String,Any}(m[1] => m[2] for m in meta)
-
-    for k in keys(metaDict)
-        val = eval(Meta.parse(metaDict[k]))
-        if isa(val[1],String)
-            val = replace.(val,Ref(" "=>""))
-        end
-        if length(val) == 1
-            val = val[1]
-        end
-        metaDict[k] = val
-    end
-    metaDict["dataprec"] = titlecase(metaDict["dataprec"])
-    return metaDict
-
-end
-
-"""
-    readAvailDiagnosticsLog(fname,fldname)
-
-Get the information for a particular field from the `available_diagnostics.log`
-    file (`MITgcm` output).
-"""
-function readAvailDiagnosticsLog(fname,fldname)
-    availdiags = readlines(fname)
-    line = availdiags[findall(occursin.(@sprintf("%-8s",fldname),availdiags))[1]]
-
-    line = split(line,'|')
-    line = lstrip.(rstrip.(line))
-
-    diagInfo = Dict([
-    "diagNum" => parse(Int,line[1]),
-    "fldname" => line[2],
-    "levs" => parse(Int,line[3]),
-    "mate" => line[4],
-    "code" => line[5],
-    "units" => line[6],
-    "title" => line[7]
-    ])
-
 end
